requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.41.0
django-crispy-forms==1.9.1
django-leaflet==0.26.0
idna==2.9
oauthlib==3.1.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
